Refining API of `DbCodec`s "extra traits" + Renames + More rustdocs + Moving codecs to separate file.
